## What changes were proposed in this pull request?

Now, we need synchronize the instance of IndexShuffleBlockResolver in order to make the commit check and tmp file rename atomically. This can be improved. We could synchronize a lock which is different for each `shuffleId + mapId` instead of  synchronize the indexShuffleBlockResolver for each writeIndexFileAndCommit call.

This should be an optimization with space for time, but it doesn't take up a lot of space.

## How was this patch tested?

Existing UT.