@cloud-fan 2018-09-10
How much perf can we save here? I don't think shuffle writing will be bottlenecked by this lock.