@cloud-fan 2018-09-11
@ConeyLiu we may have an executor lost and then come back, and may have 2 same tasks running on the same executor.